ci: adopt shipit wf-checks quality-checks caller (ADP02-WS05) (#185)
* fix: clear the clippy-1.96 backlog surfaced by the required lint lane

Making the wf-checks lint lane a required check (ADP02-WS05, #184) surfaces
the pre-existing clippy-1.96 findings the managed lint toolchain pins:
unnecessary_cast on the already-i64 `days` in standout-seeker's date/datetime
parsing, and manual_checked_ops in standout-render's tabular width resolver
(checked_div behind the existing total_weight guard — None unreachable, no
behavior change). Fixed exactly as clippy prescribes; without this the new
required lane is red on arrival.


* ci: adopt shipit wf-checks quality-checks caller (ADP02-WS05)

Cut PR-time quality checks over to shipit's wf-checks block (for #184;
epic arthur-debert/shipit#422, workstream arthur-debert/shipit#656):

- .shipit.toml [lanes]: lint + test, both required, both local. The lane
  planner (shipit ci plan) is the single decider of what runs per event.
- checks.yml: thin routing-only caller of
  arthur-debert/shipit/.github/workflows/wf-checks.yml@v1 (modeled on
  lex-fmt/lex#829) — per-event concurrency group, if: always() aggregator
  keeping the stable 'check' name alive with cancellation-skip.
- ci.yml (legacy arthur-debert/release rust-ci.yml@v3 caller) removed;
  its toolchain provisioning relocates into pixi: rust 1.96 +
  cargo-nextest + uv in [dependencies] (lanes self-provision on hosted
  runners, shipit#758), lint-full in [feature.lint.tasks] resolving into
  the lint env and provisioning lexd inline.


* render: surface guarded checked_div invariant with expect (PR #185 review)

The flex-width allocation's checked_div sits inside a total_weight > 0
guard, so the None branch is unreachable. unwrap_or(0) masked that
invariant; use .expect() so a broken guard surfaces loudly instead of
silently producing an incorrect width distribution. clippy's checked-div
lint is still satisfied by .expect().


---------

+++ crates/standout-seeker/src/parse.rs
@@ -356,7 +356,7 @@ fn parse_date_only(value: &str) -> Option<Timestamp> {
     }
 
     let days = days_from_ymd(year, month, day)?;
-    let ms = days as i64 * 24 * 60 * 60 * 1000;
+    let ms = days * 24 * 60 * 60 * 1000;
     Some(Timestamp(ms))
 }
 
@@ -395,7 +395,7 @@ fn parse_datetime(value: &str) -> Option<Timestamp> {
 
     let days = days_from_ymd(year, month, day)?;
     let seconds = hour * 3600 + minute * 60 + second;
-    let ms = days as i64 * 24 * 60 * 60 * 1000 + seconds as i64 * 1000;
+    let ms = days * 24 * 60 * 60 * 1000 + seconds as i64 * 1000;
     Some(Timestamp(ms))
 }
